+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added model to abstract away ethers types
 - [MAJOR] Updated package to es-module
 - [MAJOR] Updated onChainChanged to not refresh page if account is not initialized (to fix Phantom constant refresh)
+- [MAJOR] Refactored Web3AccountContext to use eip-6963
 
 ### Removed